# Optimization Hash calculation