Optimized Hash function , no need to create byte array,  and add method inlining to improve performance.

Before:
|      Method |      N |      Mean |     Error |    StdDev |     Gen 0 |    Gen 1 |    Gen 2 | Allocated |
|------------ |------- |----------:|----------:|----------:|----------:|---------:|---------:|----------:|
| BloomFilter | 100000 | 23.572 ms | 0.2569 ms | 0.2278 ms | 7125.0000 |        - |        - |     32 MB |
|     HashSet | 100000 |  5.528 ms | 0.0934 ms | 0.0828 ms |  554.6875 | 515.6250 | 500.0000 |      6 MB |

After:
|          Method |      N |      Mean |     Error |    StdDev |     Gen 0 |    Gen 1 |    Gen 2 | Allocated |
|---------------- |------- |----------:|----------:|----------:|----------:|---------:|---------:|----------:|
|     BloomFilter | 100000 | 21.064 ms | 0.2085 ms | 0.1848 ms | 5781.2500 |        - |        - |     26 MB |
|         HashSet | 100000 |  5.622 ms | 0.1006 ms | 0.1273 ms |  554.6875 | 500.0000 | 500.0000 |      6 MB |